requirement.txt
# requirements.txt
Django>=4.1.0,<4.2.0
djangorestframework>=3.14.0,<3.15.0
psycopg2-binary>=2.9.5,<2.10.0
django-cors-headers>=3.13.0,<3.14.0
django-allauth>=0.51.0,<0.52.0
django-phonenumber-field>=7.0.0,<7.1.0
phonenumbers>=8.13.0,<8.14.0
Pillow>=9.3.0,<9.4.0
python-dateutil>=2.8.2,<2.9.0
drf-yasg>=1.21.4,<1.22.0